BareMetal: Fix device selection properties parsing in UVSC
A problem is that the source *.pdsc file contains a five main separate
items: 'package', 'family', 'sub-family', 'device', and 'device variant'.
Each of this item may contain an unique properties, which we need to
separate in a resulting device selection structure.

But, earlier we ignore the item types at filling of the device selection,
that lead to overwriting some properties in a wrong way for a some *.pdsc
files.

To solve it, we introduce the five item types: Package, Family, SubFamily,
Device, and DeviceVariant. In this case we take in account a current item
type at filling the resulting device selection structure to exclude an
unexpected values.

Change-Id: Iaa79d1b82c284a5f30004ebdd7abc02e77872b4e
Reviewed-by: hjk <hjk@qt.io>