LSP: Fix error when opening a file with spaces in its path
When doing Follow Symbol Under Cursor to a file that has a path with
spaces, QtCreator will show a message box with an error.

The issue is that the resolved file path can still have %20 encoding
after the QUrl::toString convertion to QString.
This is because toString does not support QUrl::FullyDecoded and fallback
to PrettyDecoded instead which won't replace %20 with a space.

To fix this, QUrl::fromPercentEncoding must be used.

Also, an existing use of QUrl::fromPercentEncoding is changed to have
an utf-8 input QByteArray instead of local8Bit as fromPercentEncoding
convert %-encoded values to raw values and then convert the result
from utf-8 to unicode QString.

Change-Id: I784bcc30a67b45dcd218170854d480ebeef4237e
Reviewed-by: David Schulz <david.schulz@qt.io>