QmlDesigner: Fix navigator selection when index is invalid
This fixes the case when the navigator tree is collapsed and nothing
is selected, then a 3D object is dragged to the 3D editor. With this
fix, the tree is expanded and selection is shown. This also fixes
QDS-1892 partially (i.e. it still happens sometimes but less often).

Task-number: QDS-2024
Task-number: QDS-1892
Change-Id: If9233497d0f3c0daffafd939476d7bd64b005c79
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>