FakeVim: Treat C-S-something as "not normal" in Input::is()
The problem triggering this change is that recently, C-S-s was not
handled (as "Save all") anymore when there was nothing to change,
and the keypress ended up in FakeVim. Now FakeVim filter for _exact_
C- modifier in Input::is(), letting C-S- pass as 'S-s' i.e. starting
replacement of the rest of the line.

Change-Id: I13499f5606c28fee148e410355bec7ffd6ed116a
Reviewed-by: Christian Stenger <christian.stenger@qt.io>