LSP: Always override the assist providers if the server supports them
When the default assist provider is not DocumentContentCompletionProvider
because of some plugin that also provide an assist provider, this would
fail.

If the user has configured a language server, assume is prefer that and
always propose the server's assist providers when available.

This the case for CMake files but probably C/C++ files too if the CppEditor
plugin is configured to handle them.

Change-Id: I641d73466c8b520beb494dae8e6b274debd2881f
Reviewed-by: David Schulz <david.schulz@qt.io>