VcsOutputWindow: Fix creating links from command output
After the refactoring of the output parsers in
1c6e4fbd3211bc180, NormalMessageFormat and
ErrorMessageFormat are no longer parsed.

To allow parsing and therefore generating links,
the command output has to be declared as stdout
or stderr, depending on the severity.

Our own log and command messages are now
(in contrast to Creator 4.12) no longer parsed for
commit hashes and links, but this is acceptable.
Parsing the command output is often more interesting.

Change-Id: I7e7a96f4ff98edd88f5d9b8abddf94363768c0a6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>