BareMetal: Tune IAR compiler using 'platform codegen flags' feature
Previously we did a predefined macros generation for a some default
target CPU, which is hardcoded in a compiler.

Right now we have added the new text field in the IAR compiler settings
page, called "Platform codegen flags". In this field we can specify the
desired CPU or FPU target flags to generate an appropriate predefined
macros (e.g. such as --cpu=, --core= and so forth).

Change-Id: Ie38e077d8bac1851957085cde32661bcd189ddd2
Reviewed-by: hjk <hjk@qt.io>