QmlDesigner: Keep hasActiveTimeline QML context property up to date
Change hasActiveTimeline QML context property setting to be based on
current timeline change notifications instead of timeline node
creation, as the timeline will not be valid at the time of the node
creation. That also failed to detect timeline deletion.

Also fixed the timeline logic to clear current timeline from model
when the timeline is deleted.

Change-Id: I9191300c964bc624a958003b6aa0d5d9e77ddf9a
Fixes: QDS-2139
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>