Beautifier: Remove dependency on CppEditor and DiffEditor
Beautifier plugin depends on CppEditor and DiffEditor at the project level,
but it does not require them to be built and used.

There is also unused includes to "cppeditor/cppeditorconstants.h".
So this commit remove the unused includes (which makes Beautifier not use
any of CppEditor code) and the dependencies on CppEditor and DiffEditor.

This allow the user to disable CppEditor plugin and use LSP instead with
the basic TextEditor.

Change-Id: If83f30e754746f1635934e5f952bf9254e83c42f
Reviewed-by: Eike Ziller <eike.ziller@qt.io>