McuSupport: Enable resetting of path choosers to the default path
This change adds a "reset" button to the path choosers. Pressing it
sets the path to the default path.

The default path is now overwritable by the system settings value (if
present). This way, there is only one and only default path, which
allows us to again simplify McuPackage::writeToSettings().

Task-number: QTCREATORBUG-23860
Change-Id: I192663f3487250b9eba4773d2510abf3f9e66127
Reviewed-by: Eike Ziller <eike.ziller@qt.io>