PathListDialog: Retain original dir separators
When editing the "Path" variable on Windows, the PathListDialog
currently returns the path with all '\\' dir separators turned
into '/'. Such conversion makes sense for the "Add ..." operation
but should not be enforced for all paths.

Or in other words: Do it like EnvironmentWidget does.

This change removes conversion of the separators of all paths.

Task-number: QTCREATORBUG-24213
Change-Id: Icab1bdc190e55e8f5724a93b7cd14b08e4c68829
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>