AutoTest: Cache expanded state as well
Parsing may end up with incomplete information when parsing documents
that are currently edited and contain syntax errors due to
incomplete changes.
This resulted in not re-adding all items which in turn led to lose
also its expanded state when re-adding it in a later successful parse.
Avoid the need to expand manually again and again by caching the
expanded state for a couple of re-parses.

Change-Id: I2994b9c7ab56c0c76f416d10247e68dea271cd10
Reviewed-by: David Schulz <david.schulz@qt.io>