Target selector: Fix drawing of the run icon
With the new layout calculations (see QTCREATORBUG-24148), the
option rect can now be used to center the icon with the provided
functions in QRect.
The current code kind of works if RunColumnWidth matches the row height,
so that it can do things like subtracting vertical from horizontal
dimensions. But in fact, the icon is not centered on the option.rect.

fillRect(option.rect, Qt::red) <-to verify.

Apart from that it has a few more downsides.

This change:
- Uses the run icon made for dark background
- Paints the icon at its native 16x16px (instead of 15x15px)
- Uses QRect::center()/moveCenter() to actually center the icon

Change-Id: Ib280c650bd454b551f9e30ca108579fee06d81f0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>