QmlDesigner: Fix for broken prototypes in templates
This is just a band-aid fix.
plugins.qmltypes can contain "duplicated" cpp type declarations, some
without export, similar to a forward declaration. This is the case for
QQuickItem in templates for example.
The prototype for QQuickItem in this case is not assoicated with
QtQuick.Item, which we expect.

Task-number: QTCREATORBUG-24283
Change-Id: I3377760e29f1af5c8e948aaa1ae7883448116872
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>