QmlDesigner: Fix spinbox empty value handling
Force zero value when text value converts to NaN.

Also, onEditingFinished doesn't trigger if value is something not
accepted by the validator (the empty value), so do the equivalent
also on focus loss, if value is still dirty (meaning onEditingFinished
didn't happen). This ensures we're never left with different value
shown and written in qml file.

Change-Id: I3862d8d11adf7955f99b50b515e015ad132a5c78
Fixes: QDS-1539
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>