Debugger: disable load qml stack action for cdb
Since collecting the qml stack relies on calling functions it has a high
risk of getting stuck in that call at least with Qt 5.15. Never the less
I added the correct function in the cdbextension so when the action
get's re-enabled in the future we already start with an updated
extension.

Task-number: QTCREATORBUG-22209
Change-Id: I535f8db5a64ae48b4163d9b67e133cc3de338432
Reviewed-by: Christian Stenger <christian.stenger@qt.io>