Core: Improve handling of long output chunks
Single chunks exceeding the limit were truncated and then shown
in a single line, potentially resulting in only a few chars if there
was a newline close to the cut-off point.

Now, don't restrict to a single line in that csae.

Additionally elide in the middle, assuming this is a better compromise
than truncating at either end.

Also, make the truncation more obvious, and mention the amount of
elided characters.

Change-Id: I850e2833e7f1f8be0f584d8e4439dd1a64f851d0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>