OutputFormatter: Fix linkification of relative file paths
If we encounter a relative file path that can map to more than one
absolute file path, we do not linkify it, as that would be misleading.
However, we forgot to check whether the "different" candidates are
really different. For example, consider the following situation:
    - We have a header file /usr/include/header.h.
    - This file shows up in the compile output as "../header.h".
    - At that time, we have two search dirs /usr/include/libA and
      /usr/include/libB.
    - This resulted in two candidate file paths
      /usr/include/libA/../header.h and /usr/include/libB/../header.h
    - The relative path was rejected as ambiguous.
Fix this by checking for duplicates when gathering candidates.

Change-Id: I139b848d938113f1b5a959d8043411f7f3e809be
Reviewed-by: hjk <hjk@qt.io>