Editor: Propose full trigger match snippets
Snippets might insert a lot more text than just the completion item text
so don't hide a completion if the trigger is typed in completely. But do
not show completions when we type in a snippet with a trigger that can
be also interpreted as a keyword or ordinary completion item.

Change-Id: I25afe5f566e1c8f30621e2f82356c1c5f0af1311
Reviewed-by: Christian Stenger <christian.stenger@qt.io>